Fix CI Failures: SensitivitySelector Test and Import
