+++ .env
@@ -1,3 +1,2 @@
 GEMINI_API_KEY=YOUR_GEMINI_API_KEY
 DEEPGRAM_API_KEY=YOUR_DEEPGRAM_API_KEY
-REACT_APP_DEEPGRAM_API_KEY=YOUR_DEEPGRAM_API_KEY
