+++ package.json
@@ -22,7 +22,6 @@
     "cors": "^2.8.5",
     "dotenv": "^17.2.3",
     "express": "^5.2.1",
-    "express-rate-limit": "^8.2.1",
     "gemini-mcp-tool": "^1.1.4",
     "gemini-zod": "^0.1.2",
     "jsonwebtoken": "^9.0.3",
@@ -59,6 +58,7 @@
   },
   "overrides": {
     "nth-check": "^2.0.1",
-    "postcss": "^8.4.31"
+    "postcss": "^8.4.31",
+    "webpack-dev-server": "^5.2.0"
   }
 }
