+++ src/nlp/cd-detector.js
@@ -54,12 +54,7 @@ async function generateReferenceEmbeddings() {
 
 // Function to calculate cosine similarity between two tensors
 function cosineSimilarity(vec1, vec2) {
-  return tf.tidy(() => {
-    const dotProduct = tf.sum(tf.mul(vec1, vec2));
-    const norm1 = tf.norm(vec1);
-    const norm2 = tf.norm(vec2);
-    return dotProduct.div(tf.mul(norm1, norm2));
-  });
+  return tf.metrics.cosineDistance(vec1, vec2).neg().add(1);
 }
 
 // sensitivity: 'low', 'medium', 'high'
